refactor: gathers all context annotations into one package
