+++ core/src/main/java/tech/guilhermekaua/spigotboot/core/context/annotations/Component.java
@@ -20,7 +20,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package tech.guilhermekaua.spigotboot.core.module.annotations;
+package tech.guilhermekaua.spigotboot.core.context.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -28,9 +28,7 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-public @interface ConditionalOnClass {
-    Class<?>[] value();
+@Target({ElementType.TYPE})
 
-    String message() default "";
+public @interface Component {
 }
