+++ core/src/main/java/tech/guilhermekaua/spigotboot/core/context/annotations/Service.java
@@ -20,14 +20,15 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package tech.guilhermekaua.spigotboot.core.listener.annotations;
+package tech.guilhermekaua.spigotboot.core.context.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+@Component
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-public @interface ListenerRegister {
+@Target({ElementType.TYPE, ElementType.METHOD})
+public @interface Service {
 }
