+++ core/src/main/java/tech/guilhermekaua/spigotboot/core/listener/manager/ListenerManager.java
@@ -27,8 +27,8 @@ import org.bukkit.event.HandlerList;
 import org.bukkit.event.Listener;
 import org.jetbrains.annotations.NotNull;
 import tech.guilhermekaua.spigotboot.core.ApxPlugin;
+import tech.guilhermekaua.spigotboot.core.context.annotations.ListenerRegister;
 import tech.guilhermekaua.spigotboot.core.di.manager.DependencyManager;
-import tech.guilhermekaua.spigotboot.core.listener.annotations.ListenerRegister;
 import tech.guilhermekaua.spigotboot.core.utils.ReflectionUtils;
 import tech.guilhermekaua.spigotboot.utils.ProxyUtils;
