+++ modules/data-orm-lite/src/main/java/tech/guilhermekaua/spigotboot/data/ormLite/registry/OrmLiteRepositoryRegistry.java
@@ -27,7 +27,7 @@ import com.j256.ormlite.dao.DaoManager;
 import com.j256.ormlite.support.ConnectionSource;
 import lombok.RequiredArgsConstructor;
 import org.bukkit.plugin.Plugin;
-import tech.guilhermekaua.spigotboot.core.di.annotations.Component;
+import tech.guilhermekaua.spigotboot.core.context.annotations.Component;
 import tech.guilhermekaua.spigotboot.core.di.manager.DependencyManager;
 import tech.guilhermekaua.spigotboot.core.reflection.DiscoveryService;
 import tech.guilhermekaua.spigotboot.core.utils.ReflectionUtils;
